feat: implement SQS message dumping and import notification workflow with error handling and progress reporting
